result estimate fee null also if the error is inside the request
I discovered that sometimes esplora retrun an error inside the request but the cause is the feerate unavailable.

For this reason, I choose to return the null answer also in cases of request error, if the error is the connection in this cases there are other method that can fail, eg. getutxo.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>