Return a failure in case the sendrawtx method return an error.
With the actual version lightnind for debugging it throws in the blockchain an transaction that is already inside the blockchain.
However, when lightnind (or other implementations) close a channel and both of the part agree, a concurrence event can happen on the network because both of the node put in the blockchain the same transaction.

p.s: The real motivation can be "A return is missing in case of the request on esplora will fail".

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>